requirements.txt
pyyaml
pymongo
dash
pandas
plotly